requirements.txt
autopep8==1.5.6
Django==2.2.19
django-cors-headers==3.7.0
djangorestframework==3.12.4
importlib-metadata==3.9.0
Markdown==3.3.4
pycodestyle==2.7.0
PyMySQL==1.0.2
pytz==2021.1
PyYAML==5.4.1
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.7.4.3
uritemplate==3.0.1
zipp==3.4.1
